+++ core/host.py
@@ -2,8 +2,17 @@ import os
 import subprocess
 import requests
 from urllib.parse import urlparse
-from core.config import DEPENDENCIES_SCRIPT, URLS, SKALE_NODE_UI_PORT, DEFAULT_URL_SCHEME
+
+from core.resources import save_resource_allocation_config
+
+from core.config import DEPENDENCIES_SCRIPT, URLS, SKALE_NODE_UI_PORT, DEFAULT_URL_SCHEME, \
+    INSTALL_CONVOY_SCRIPT
+from configs.node import NODE_DATA_PATH
+from configs.resource_allocation import DISK_MOUNTPOINT_FILEPATH, \
+    CONVOY_HELPER_SCRIPT_FILEPATH, CONVOY_SERVICE_TEMPLATE_PATH, CONVOY_SERVICE_PATH
+
 from core.helper import safe_get_config, safe_load_texts, construct_url, clean_cookies, clean_host
+from tools.helper import run_cmd, process_template, get_username
 
 TEXTS = safe_load_texts()
 
@@ -54,3 +63,43 @@ def fix_url(url):
         return url
     except ValueError:
         return False
+
+
+def prepare_host(test_mode, disk_mountpoint):
+    init_data_dir()
+    save_disk_mountpoint(disk_mountpoint)
+    save_resource_allocation_config()
+    if not test_mode:
+        init_convoy(disk_mountpoint)
+
+def init_convoy(disk_mountpoint):
+    print(f'Installing convoy...')
+    run_cmd(['bash', INSTALL_CONVOY_SCRIPT], shell=False)
+    print(f'Downloading convoy disk helper...')
+    convoy_prepare_disk(disk_mountpoint)
+    start_convoy_daemon(disk_mountpoint)
+
+
+def start_convoy_daemon(disk_mountpoint):
+    template_data = {
+        'user': get_username(),
+        'cmd': f'/usr/local/bin/convoy daemon --drivers devicemapper --driver-opts dm.datadev={disk_mountpoint}1 --driver-opts dm.metadatadev={disk_mountpoint}2'
+    }
+    process_template(CONVOY_SERVICE_TEMPLATE_PATH, CONVOY_SERVICE_PATH, template_data)
+    run_cmd(['systemctl', 'start', 'convoy'], shell=False)
+
+
+def convoy_prepare_disk(disk_mountpoint):
+    print(f'Applying disk partitioning...')
+    run_cmd(['bash', CONVOY_HELPER_SCRIPT_FILEPATH, '--write-to-disk', f'{disk_mountpoint}'],
+            shell=False)
+
+
+def save_disk_mountpoint(disk_mountpoint):
+    with open(DISK_MOUNTPOINT_FILEPATH, 'w') as f:
+        f.write(disk_mountpoint)
+
+
+def init_data_dir():
+    print(f'Creating {NODE_DATA_PATH} directory...')
+    os.makedirs(NODE_DATA_PATH, exist_ok=True)