+++ datafiles/install.sh
@@ -7,6 +7,8 @@ cd /skale
 sudo git clone -b $GIT_BRANCH https://$GITHUB_TOKEN\@github.com/skalenetwork/skale-node.git
 #cp -f /tmp/data.json ./skale-node/config/data.json
 
+umount $DISK_MOUNTPOINT
+
 cd /skale/skale-node/installation
 DOCKER_USERNAME=$DOCKER_USERNAME DOCKER_PASSWORD=$DOCKER_PASSWORD RUN_MODE=prod RPC_IP=$RPC_IP \
     RPC_PORT=$RPC_PORT DB_USER=$DB_USER DB_PASSWORD=$DB_PASSWORD DB_ROOT_PASSWORD=$DB_ROOT_PASSWORD \