+++ tools/schain_types.py
@@ -3,6 +3,7 @@ from enum import Enum
 
 class SchainTypes(Enum):
     test = 0
+    test4 = 4
     large = 1
     medium = 8
     small = 128