+++ CHANGELOG.md
@@ -4,7 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased, [0.1.0.dev.0]
+## Unreleased
+
+### Added 
+
+- Logger module
+- Logs for most CLI calls
+- `skale logs cli` command to fetch CLI logs, `--debug` option for more detailed output
+
+### Changed
+
+### Fixed
+
+
+## [0.1.0] - 2019-07-31
 
 > Breaking changes, no backward compatibility with 0.0.x releases!